Allow to set post_method in HttpPostClientTransport
Fixes #26.

In some applications it is desirable to limit the number of ports used by a
client (see issue #26). This change allows to supply an alternative method to
the default `requests.post` being used in `send_message`.

The change is introduced backwards compatible with the introduction of a new
kwarg to the constructor, `post_method=None`.

Users can create a `requests.Session()` instance, (optionally mount a
`requests.adapters.HTTPAdapter` to control the pool size), and supply
`post_method=session.post` as the kwarg.

`test_wsgi_transport.py::sessioned_client` shows an application of the kwarg.
`test_wsgi_transport.py::test_exhaust_ports` displays the problem with a non
sesioned `HttpPostClientTransport` instance.